requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_py-xgboost-mutex=2.0=cpu_0
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py37_0
alabaster=0.7.12=py37_0
anykeystore=0.2=pypi_0
apex=0.9.10.dev0=pypi_0
argh=0.26.2=py37_0
asn1crypto=1.2.0=py37_0
astor=0.8.0=py37_0
astroid=2.3.3=py37_0
atomicwrites=1.3.0=py37_1
attrs=19.3.0=py_0
autoflake=1.3.1=pypi_0
autopep8=1.4.4=py_0
babel=2.7.0=py_0
backcall=0.1.0=py37_0
beautifulsoup4=4.8.2=pypi_0
biopython=1.76=pypi_0
blas=1.0=mkl
bleach=3.1.0=py37_0
boto3=1.10.44=pypi_0
botocore=1.13.44=pypi_0
bs4=0.0.1=pypi_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.11.28=hecc5488_0
cachetools=4.0.0=pypi_0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=py37_1003
cloudpickle=1.2.2=py_0
colorama=0.4.3=pypi_0
cryptacular=1.5.5=pypi_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.15.2=py37_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
feedparser=5.2.1=py37_1
flake8=3.7.9=py37_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=py37_0
gast=0.2.2=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
google-auth=1.10.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py37hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py37h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
hupper=1.9.1=pypi_0
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imagesize=1.1.0=py37_0
importlib_metadata=1.3.0=py37_0
intel-openmp=2019.5=281
intervaltree=3.0.2=py_0
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.10.2=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py37_0
jedi=0.14.1=py37_0
jeepney=0.4.1=py_0
jinja2=2.10.3=py_0
jmespath=0.9.4=pypi_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
json5=0.8.5=py_0
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=5.2.0=py37_1
jupyter_core=4.6.1=py37_0
jupyterlab=1.2.4=py_0
jupyterlab_server=1.0.6=py_0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-gpu=2.2.4=0
keras-preprocessing=1.1.0=py_1
keras-tuner=1.0.0=pypi_0
keyring=20.0.0=py37_0
kiwisolver=1.1.0=py37he6710b0_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=h4deb6c0_3
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.2=hd408876_0
libsodium=1.0.16=h1bed415_0
libspatialindex=1.9.3=he6710b0_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxgboost=0.90=he1b5a44_4
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.1=py37h5429711_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2019.5=281
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
more-itertools=8.0.2=py_0
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.5=py37_0
notebook=6.0.1=py37_0
numpy=1.17.4=py37hc1035e2_0
numpy-base=1.17.4=py37hde5b4d6_0
numpydoc=0.9.1=py_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1=h7b6447c_0
opt_einsum=3.1.0=py_0
packaging=19.2=py_0
pandas=0.25.3=py37he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.2=py_0
pastedeploy=2.0.1=pypi_0
pathtools=0.1.2=py_1
patsy=0.5.1=py37_0
pbkdf2=1.3=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pip=19.3.1=py37_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
pluggy=0.13.1=py37_0
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.2=py_0
protobuf=3.11.2=py37he6710b0_0
psutil=5.6.7=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
py-xgboost=0.90=py37_4
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py37_0
pygments=2.5.2=py_0
pylint=2.4.4=py37_0
pympler=0.7=py_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h05f1152_2
pyramid=1.10.4=pypi_0
pyramid-mailer=0.15.1=pypi_0
pyrsistent=0.15.6=py37h7b6447c_0
pysocks=1.7.1=py37_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.2=py_0
python-language-server=0.31.2=py37_0
python3-openid=3.1.0=pypi_0
pytorch=1.3.1=py3.7_cuda10.0.130_cudnn7.6.3_0
pytorch-transformers=1.2.0=pypi_0
pytz=2019.3=py_0
pyxdg=0.26=py_0
pyyaml=5.2=py37h7b6447c_0
pyzmq=18.1.0=py37he6710b0_0
qdarkstyle=2.7=py_0
qt=5.9.7=h5867ecd_1
qtawesome=0.6.0=py_0
qtconsole=4.6.0=py_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
regex=2019.12.20=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.22.0=py37_1
requests-oauthlib=1.3.0=pypi_0
rope=0.14.0=py_0
rsa=4.0=pypi_0
rtree=0.8.3=py37_0
s3transfer=0.2.1=pypi_0
sacremoses=0.0.35=pypi_0
scikit-learn=0.22=py37hd81dba3_0
scipy=1.3.2=py37h7c811a0_0
seaborn=0.9.0=pyh91ea838_1
secretstorage=3.1.1=py37_0
send2trash=1.5.0=py_0
sentencepiece=0.1.85=pypi_0
seqeval=0.0.12=pypi_0
setuptools=42.0.2=py37_0
simpletransformers=0.13.4=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.1.0=py37_0
soupsieve=1.9.5=pypi_0
sphinx=2.3.0=py_0
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
spyder=4.0.0=py37_0
spyder-kernels=1.8.1=py37_0
spyder-notebook=0.2.1=py_0
sqlalchemy=1.3.12=pypi_0
sqlite=3.30.1=h7b6447c_0
statsmodels=0.10.1=py37hdd07704_0
tabulate=0.8.6=pypi_0
tensorboard=2.1.0=pypi_0
tensorboardx=1.9=pypi_0
tensorflow=2.1.0rc2=pypi_0
tensorflow-base=2.0.0=gpu_py37h0ec5d1f_0
tensorflow-estimator=2.1.0=pypi_0
tensorflow-gpu=2.0.0=h0d30ee6_0
tensorflow-hub=0.7.0=pyhe1b5a44_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
terminaltables=3.1.0=pypi_0
testpath=0.4.4=py_0
textblob=0.15.3=py_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.41.0=pypi_0
traitlets=4.3.3=py37_0
transaction=3.0.0=pypi_0
transformers=2.3.0=pypi_0
translationstring=1.3=pypi_0
ujson=1.35=py37h14c3975_0
urllib3=1.25.7=py37_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
watchdog=0.9.0=py37_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
webob=1.8.5=pypi_0
werkzeug=0.16.0=py_0
wget=3.2=pypi_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.11.2=py37h7b6447c_0
wtforms=2.2.1=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
wurlitzer=2.0.0=py37_0
xgboost=0.90=py37he1b5a44_4
xlrd=1.2.0=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=h96e3832_1
yapf=0.28.0=py_0
zeromq=4.3.1=he6710b0_3
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zope-deprecation=4.4.0=pypi_0
zope-interface=4.7.1=pypi_0
zope-sqlalchemy=1.2=pypi_0
